Introduce flowReset and listReset with Arrow support
