+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 agp = "8.0.2"
-arrow = "1.2.4"
+arrow = "2.0.0-SNAPSHOT"
 kotestAssertionsCore = "5.8.1"
 kotlin = "2.0.0-RC1"
 kotlinxCoroutinesCore = "1.8.0"
