+++ library/src/commonTest/kotlin/ListTest.kt
@@ -1,3 +1,4 @@
+import arrow.core.raise.ensure
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
@@ -7,12 +8,12 @@ class ListTest {
   fun empty() = runTest {
     val list = emptyList<Int>()
     var counter = 0
-    val result = reset<List<Int>> {
+    val result = listReset {
       val item = list.bind()
       effect {
         counter++
       }
-      listOf(item)
+      item
     }
     result shouldBe list
     counter shouldBe list.size
@@ -22,12 +23,12 @@ class ListTest {
   fun single() = runTest {
     val list = listOf(1, 2, 3)
     var counter = 0
-    val result = reset<List<Int>> {
+    val result = listReset {
       val item = list.bind()
       effect {
         counter++
       }
-      listOf(item)
+      item
     }
     result shouldBe list
     counter shouldBe list.size
@@ -36,10 +37,11 @@ class ListTest {
   @Test
   fun filtering() = runTest {
     val list = listOf(1, 2, 3)
-    val result = reset<List<Int>> {
+    val result = listReset {
       val item = list.bind()
       effect {
-        if (item == 2) emptyList() else listOf(item)
+        ensure(item != 2) { }
+        item
       }
     }
     result shouldBe list.filter { it != 2 }
@@ -54,25 +56,25 @@ class ListTest {
     var firstCounter = 0
     var secondCounter = 0
     var thirdCounter = 0
-    val result = reset<List<Pair<Int, Int>>> {
+    val result = listReset {
       effect {
         noObservedCounter++
       }
       val first = list1.bind()
       effect {
-        if (first == Int.MAX_VALUE) return@reset emptyList()
+        ensure(first != Int.MAX_VALUE) { }
         firstCounter++
       }
       val second = list2.bind()
       effect {
-        if (second == Int.MAX_VALUE) return@reset emptyList()
+        ensure(second != Int.MAX_VALUE) { }
         secondCounter++
       }
       list3.bind()
       effect {
         thirdCounter++
       }
-      listOf(first to second)
+      first to second
     }
     result shouldBe list1.filter { it != Int.MAX_VALUE }.flatMap { first ->
       list2.filter { it != Int.MAX_VALUE }.flatMap { second ->
@@ -92,7 +94,7 @@ class ListTest {
     val list = listOf(listOf(1, 2), listOf(3, 4), listOf(5, 6))
     var innerCount = 0
     var itemCount = 0
-    val result = reset<List<Int>> {
+    val result = listReset {
       val inner = list.bind()
       effect {
         innerCount++
@@ -101,7 +103,7 @@ class ListTest {
       effect {
         itemCount++
       }
-      listOf(item)
+      item
     }
     result shouldBe list.flatten()
     innerCount shouldBe list.size
@@ -112,16 +114,16 @@ class ListTest {
   fun ifElse() = runTest {
     val list = listOf(1, 2, 2, 3)
     val twoElements = listOf(0, 0)
-    val result = reset<List<String>> {
+    val result = listReset {
       val x = listOf(1, 2, 2, 3).bind()
       if (x == 2) {
         twoElements.bind()
-        listOf("firstBranch")
+        "firstBranch"
       } else {
         repeat(2) {
           twoElements.bind()
         }
-        listOf("secondBranch")
+        "secondBranch"
       }
     }
     result shouldBe list.flatMap {
@@ -133,11 +135,11 @@ class ListTest {
 
   @Test
   fun forLoops() = runTest {
-    val result = reset<List<Int>> {
+    val result = listReset {
       for (i in 1..10) {
         listOf(i, i).bind()
       }
-      listOf(0)
+      0
     }
     result shouldBe List(1024) { 0 }
   }
